+++ packages/ListBox/src/index.d.ts
@@ -142,6 +142,12 @@ declare namespace ListBox {
     label?: string;
     /** Callback for the clicking event */
     onClick?: (...args: any[]) => any;
+
+    handleOnClick: (...args: any[]) => any;
+
+    id?: string;
+
+    isSelectedValue: boolean;
     /** Value of your option this can be any data structure */
     value?: any;
     /** Internal prop, which shouldn't be documented */
