+++ packages/ListBox/stories/ListBox.backyard.stories.js
@@ -4,6 +4,7 @@ import { exampleStoryParameters } from "storybook/assets/storyParameters";
 import ExampleStory from "storybook/components/ExampleStory";
 import DirectoryExample from "./sandbox/Directory/Directory";
 import FilterSelectExample from "./sandbox/FilterSelect/FilterSelect";
+import LongListExample from "./sandbox/LongList/LongList";
 import CollapsedContainerExample from "./sandbox/CollapsedContainer";
 import Lazy from "./sandbox/LazyListBox/Lazy";
 import ListBox from "../src";
@@ -13,6 +14,19 @@ const storyName = getStoryName("ListBox");
 export default {
   title: `${storyName}/Backyard/Sandbox`,
   component: ListBox,
+};
+
+export const LongListStory = () => (
+  <ExampleStory storyName="Long List" component="ListBox" fileName="sandbox/LongList/LongList.js">
+    <LongListExample />
+  </ExampleStory>
+);
+
+LongListStory.story = {
+  name: "Long List",
+  parameters: {
+    ...exampleStoryParameters,
+  },
 };
 
 export const FilterSelectStory = () => (
