+++ packages/ListBox/stories/ListBox.single.stories.js
@@ -2,7 +2,6 @@ import React from "react";
 import { getStoryName } from "storybook/storyTree";
 import ExampleStory from "storybook/components/ExampleStory";
 import { Controlled } from "./examples/Single/Controlled";
-import { Performance } from "./examples/Single/Performance";
 import { Inline } from "./examples/Single/Inline";
 import ListBox from "../src";
 
@@ -38,15 +37,5 @@ export const InlineStory = () => (
 );
 InlineStory.story = {
   name: "Inline",
-  parameters: paramaters,
-};
-
-export const PerformanceStory = () => (
-  <ExampleStory component="ListBox" storyName="Performance ListBox" fileName="examples/Single/Performance.js">
-    <Performance />
-  </ExampleStory>
-);
-ControlledStory.story = {
-  name: "Controlled",
   parameters: paramaters,
 };
